requirements.txt
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.100.0
requests>=2.31.0
python-dotenv>=1.0.0
